requirements.txt
langchain
langchain_community
faiss-cpu
pypdf
sentence-transformers
beautifulsoup4
from_root
langchain-huggingface
-e .